Fix fitsIn function and move it's test to tests folder
